Update road-runner-gui.md (#82)
## Description of issue
The Tools > Official Road Runner GUI page had a link to a jar released in 2020. 

## Description of change
I changed the link to use the `latest` tag in the repo which will redirect to the latest stable release.

I also added an example command to run the jar for those who are new to Java.